refactor: domain/history の型ファイルを types.ts に集約
5 ファイルに分けていた型定義 (HistoryItem, HistoryStore, ParsedSearchQuery,
GroupingOptions, SearchOptions) はそれぞれ小さく、関連も強いため
単一ファイル types.ts にまとめて読みやすさを優先する。

「関数は分離 / 型は集約」の方針:
- 関数は実装の塊として独立した責務を持つので 1 ファイル 1 関数
- 型は宣言だけで小さく、同ドメインで関連が強いので 1 ファイルに集約
