+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 
 import { Root } from "./components/history/Root";
 import { searchHistories } from "./domain/history/searchHistories";
-import { HistoryItem } from "./domain/history/types/HistoryItem";
+import { HistoryItem } from "./domain/history/types";
 import { t } from "./i18n";
 import { bookmarkHistoryStore } from "./infra/bookmark-history-store";
 import {
