+++ src/components/history/HistoryDropdown/index.tsx
@@ -1,6 +1,6 @@
 import { FC, useMemo } from "react";
 
-import { HistoryItem } from "../../../domain/history/types/HistoryItem";
+import { HistoryItem } from "../../../domain/history/types";
 import { t } from "../../../i18n";
 import { Dropdown } from "../../common/Dropdown";
