--- src/domain/history/types/GroupingOptions.ts
@@ -1,4 +1,0 @@
-export interface GroupingOptions {
-  groupByUrl?: boolean;
-  groupByTitle?: boolean;
-}
