--- src/domain/history/types/HistoryItem.ts
@@ -1,9 +1,0 @@
-export type HistoryItem = {
-  id: string;
-  url: string;
-  title: string;
-  visitCount: number;
-  lastVisitTime: number;
-  domain: string;
-  favicon?: string;
-};
