--- src/domain/history/types/ParsedSearchQuery.ts
@@ -1,4 +1,0 @@
-export type ParsedSearchQuery = {
-  term: string;
-  type: "text" | "site" | "exclude";
-};
